Fixed error with wrong videos showing up
