+++ src/App.tsx
@@ -3,8 +3,8 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import AuthPage from "./AuthPage";
 import HomePage from "./home-page";
 import VideoOne from "./course-listing-pages/password-security/video1";
-import VideoTwo from "./course-listing-pages/phishing-and-social-engineering/video3";
-import VideoThree from "./course-listing-pages/physical-security/video2";
+import VideoThree from "./course-listing-pages/phishing-and-social-engineering/video3";
+import VideoTwo from "./course-listing-pages/physical-security/video2";
 import PrivateRoute from "./PrivateRoute";
 import CourseListing1 from "./course-listing-pages/password-security/course-listing1";
 import CourseListing2 from "./course-listing-pages/physical-security/course-listing2";
